lava-stream-job: move get_config into lib/configuration
